Moving `PostEditorNavigationBarManager` to its own file.
